+++ packages/hot-fuzz/default.nix
@@ -1,0 +1,17 @@
+{
+  halfBoardModule = {
+    dependencies = [ ../ethereum-contracts];
+    outputs = [
+      "build"
+      "crytic-export"
+    ];
+    includedFiles = [
+      ./package.json
+      ./foundry.toml
+      ./contracts
+      ./scripts
+      ./hot-fuzz
+      ./Makefile
+    ];
+  };
+}